automation_service/requirements.txt
requests
woocommerce
streamlit
beautifulsoup4
python-dotenv
openai
fastapi
uvicorn

**
Starting a **barcode (UPC/EAN/GTIN) database** is very doable, but it helps to be clear about **what kind of database you want** and **how official you want it to be**. Below is a practical, step-by-step guide—from concept to implementation—based on how existing barcode databases work.

---

## 1. Decide the scope (this is critical)

First, define **what your database will and will not do**.

### Common models

**A. Product lookup database (most common)**

* Input: UPC / EAN / GTIN
* Output: product name, brand, category, image, description
* Examples: Go-UPC, Barcode Spider

**B. Verification database**

* Input: barcode number
* Output: valid/invalid, GS1 prefix, issuing country
* No product images or descriptions

**C. Regional or niche database**

* Example: *Jamaica-only products*, Caribbean products, local manufacturers
* Much easier to start and grow

👉 **Recommendation:** Start with **C**, then expand.

---

## 2. Understand the legal reality (important)

* **GS1 owns the standard, not the data**
* You **cannot issue “official” GTINs** unless you are GS1
* You **can store, index, and display barcode → product data**

✅ Legal:

* Storing GTINs
* Allowing manufacturers to submit product info
* Showing public product data

❌ Not allowed:

* Claiming your codes are GS1-issued
* Selling “official” barcodes

---

## 3. Data model (what you actually store)

A simple but solid schema:

### Core tables

**Products**

* `gtin` (primary key)
* `brand`
* `product_name`
* `category`
* `description`
* `image_url`
* `status` (active / discontinued)

**Companies**

* `company_id`
* `company_name`
* `country`
* `gs1_prefix` (optional)

**Metadata**

* `gtin`
* `created_at`
* `updated_at`
* `source` (manufacturer, retailer, public)

---

## 4. Where the data comes from (hardest part)

### Option 1: Manufacturer self-submission (best)

* Businesses create accounts
* They submit their own GTINs and product info
* You review/approve submissions

✔ Most accurate
✔ Scales well
✖ Slow at first

---

### Option 2: Public contributions (Wikipedia model)

* Anyone can submit or edit
* Moderation system required

✔ Faster growth
✖ Quality control needed

---

### Option 3: Partnerships

* Retailers
* Distributors
* Local trade organizations

✔ High-quality data
✔ Regional strength
✖ Requires outreach

---

### ⚠️ Web scraping

* Legally risky
* Often violates terms of service
* Not recommended for long-term projects

---

## 5. Barcode validation logic (must-have)

Even without product data, you should:

* Validate **check digits** (EAN-13, UPC-A)
* Identify **barcode type**
* Identify **GS1 prefix (country/region)**

This builds trust and credibility.

---

## 6. Tech stack (simple & proven)

### Backend

* **PostgreSQL** or **MySQL**
* **Node.js / Python (FastAPI or Django)**
* REST or GraphQL API

### Frontend

* Simple search page (barcode input)
* Product result page
* Admin dashboard

### Optional

* Barcode scanning (mobile)
* Public API (rate-limited)

---

## 7. Trust & credibility features

To be taken seriously:

* Source labels (“Manufacturer-submitted”)
* Verification badges
* Change history
* Contact & correction requests

---

## 8. Monetization (optional)

Most successful databases use **freemium** models:

* Free basic lookup
* Paid API access
* Paid bulk queries
* Business dashboards
* Product verification badges

---

## 9. Example MVP roadmap (realistic)

**Month 1**

* Barcode validation
* Search interface
* Manual product entry

**Month 2**

* Manufacturer accounts
* Submission workflow
* Moderation tools

**Month 3**

* API
* Public search indexing
* Regional marketing

---

## 10. If your goal is Jamaica / Caribbean (very smart niche)

You could:

* Partner with **local manufacturers**
* Work with **retail associations**
* Become the **de facto regional product registry**
* Later integrate with GS1 services (not compete with them)

---

### Bottom line

You’re not building a barcode system—you’re building a **product identity database**.

If you want, tell me:

* 🌍 **Region** (Jamaica only? Caribbean? global?)
* 👥 **Target users** (manufacturers, consumers, retailers, developers)
* 🛠 **Technical skill level**

…and I’ll help you design a **concrete architecture + launch plan** tailored exactly to that.

**